[IMP] Unified evidences modal liek the rest of the case objects
