+++ source/app/blueprints/case/templates/case_rfile.html
@@ -28,7 +28,7 @@
                             </button>
                         </li>
                         <li class="nav-item">
-                            <button data-toggle="modal"  data-target="#modal_add_receivedfile" class="btn btn-dark btn-sm">
+                            <button onclick="add_modal_rfile()" class="btn btn-dark btn-sm">
                                  <span class="menu-title">Register Evidence</span>
                             </button>
                         </li>
@@ -87,6 +87,7 @@
             </div>
             <!-- /.modal-dialog -->
         </div>
+
         {% endif %}
     </div>
