+++ source/app/templates/includes/quick-sidebar.html
@@ -55,13 +55,13 @@
 								<div class="quick-actions-scroll scrollbar-outer">
 									<div class="quick-actions-items">
 										<div class="row m-0">
-											<a class="col-6 col-md-6 p-0" href="/case/ioc">
+											<a class="col-6 col-md-6 p-0" href="/case/ioc?cid={{case_name[2]}}">
 												<div class="quick-actions-item">
 													<i class="flaticon-inbox"></i>
 													<span class="text">Add IOC</span>
 												</div>
 											</a>
-											<a class="col-6 col-md-6 p-0" href="/case/assets">
+											<a class="col-6 col-md-6 p-0" href="/case/assets?cid={{case_name[2]}}">
 												<div class="quick-actions-item">
 													<i class="flaticon-laptop"></i>
 													<span class="text">Add asset</span>
@@ -69,13 +69,13 @@
 											</a>
 										</div>
 										<div class="row m-0">
-											<a class="col-6 col-md-6 p-0" href="/case/notes">
+											<a class="col-6 col-md-6 p-0" href="/case/notes?cid={{case_name[2]}}">
 												<div class="quick-actions-item">
 													<i class="flaticon-file-1"></i>
 													<span class="text">Add note</span>
 												</div>
 											</a>
-											<a class="col-6 col-md-6 p-0" href="/case/timeline">
+											<a class="col-6 col-md-6 p-0" href="/case/timeline?cid={{case_name[2]}}">
 												<div class="quick-actions-item">
 													<i class="flaticon-stopwatch"></i>
 													<span class="text">Add event</span>
@@ -83,7 +83,7 @@
 											</a>
 										</div>
                                         <div class="row m-0">
-											<a class="col-6 col-md-6 p-0" href="/manage/cases">
+											<a class="col-6 col-md-6 p-0" href="/manage/cases?cid={{case_name[2]}}">
 												<div class="quick-actions-item">
 													<i class="flaticon-box"></i>
 													<span class="text">Upload evidence</span>
@@ -97,7 +97,7 @@
 											</a>
 										</div>
                                         <div class="row m-0">
-											<a class="col-6 col-md-6 p-0" data-toggle="modal" data-target="#modal_add_receivedfile" href="#">
+											<a class="col-6 col-md-6 p-0" href="/case/evidences?cid={{case_name[2]}}">
 												<div class="quick-actions-item">
 													<i class="flaticon-price-tag"></i>
 													<span class="text">Register evidence</span>
